+++ setup.sh
@@ -3,21 +3,21 @@
 CHROME_URI="https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb"
 
 for i in apt curl pip python; do
-  command -v "$i" >/dev/null || {
+  command -v "$i" > /dev/null || {
     echo "require: $i" >&2
     exit 1
   }
 done
 
 echo "[install google-chrome]"
-command -v google-chrome >/dev/null || {
+command -v google-chrome > /dev/null || {
   out="/tmp/google-chrome-stable.deb"
   [ -f "$out" ] || curl -s "$CHROME_URI" \
     -o "$out" \
     sudo apt -f install -y -qq "$out"
 }
 
-command -v google-chrome >/dev/null || {
+command -v google-chrome > /dev/null || {
   echo "failed to install: google-chrome" >&2
   exit 1
 }
@@ -32,10 +32,10 @@ dpkg -l | grep python3-selenium -q || {
 }
 
 echo "[install selenium module to python]"
-python -c "import selenium;print(selenium.__version__)" >/dev/null || {
+python -c "import selenium;print(selenium.__version__)" > /dev/null || {
   python -m pip install selenium
 }
-python -c "import selenium;print(selenium.__version__)" >/dev/null || {
+python -c "import selenium;print(selenium.__version__)" > /dev/null || {
   echo "failed to install: selenium module" >&2
   exit 1
 }