+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## [v1.5](https://github.com/eggplants/get-pixivpy-token/releases/tag/v1.5) - 2021-09-15 19:14:48 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v1.4...v1.5))
+
+- [24d635dd5eea2530d03413469524c72910ef228a](https://github.com/eggplants/get-pixivpy-token/commit/24d635dd5eea2530d03413469524c72910ef228a)
+  - v1.5
+- [11f5dc0c5fed7b72c718c6feb307dcb2d08271a9](https://github.com/eggplants/get-pixivpy-token/commit/11f5dc0c5fed7b72c718c6feb307dcb2d08271a9)
+  - fix: login method (#3)
+
+## [v1.4](https://github.com/eggplants/get-pixivpy-token/releases/tag/v1.4) - 2021-09-08 01:51:49 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v1.2...v1.4))
+
+- [38401c27d54737097b4adae006bc778709c10120](https://github.com/eggplants/get-pixivpy-token/commit/38401c27d54737097b4adae006bc778709c10120)
+  - v1.4
+- [0a660988b3642d87004c3fd550c8ee53cf677885](https://github.com/eggplants/get-pixivpy-token/commit/0a660988b3642d87004c3fd550c8ee53cf677885)
+  - v1.3: add interactive mode
+
+## [v1.2](https://github.com/eggplants/get-pixivpy-token/releases/tag/v1.2) - 2021-09-08 00:09:46 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v1.1...v1.2))
+
+- [6b1c51b8cf6916814d5a87b4095cd8bf054f175f](https://github.com/eggplants/get-pixivpy-token/commit/6b1c51b8cf6916814d5a87b4095cd8bf054f175f)
+  - v1.2
+- [367cdead110071cdc36b043383bd72cfe1b2e542](https://github.com/eggplants/get-pixivpy-token/commit/367cdead110071cdc36b043383bd72cfe1b2e542)
+  - fix: change location of auth info from init method into login method
+
+## [v1.1](https://github.com/eggplants/get-pixivpy-token/releases/tag/v1.1) - 2021-09-06 15:29:25 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v1.0...v1.1))
+
+- [57d1e9de11a89f7d30656085baa248a99757cb8f](https://github.com/eggplants/get-pixivpy-token/commit/57d1e9de11a89f7d30656085baa248a99757cb8f)
+  - v1.1
+- [d3cd297bc8711a0f16599f422eba58e537da9e6e](https://github.com/eggplants/get-pixivpy-token/commit/d3cd297bc8711a0f16599f422eba58e537da9e6e)
+  - add: make importimg GetPixivToken easier
+- [bccbc4f7262855f034b881dc303b0e7488d813fb](https://github.com/eggplants/get-pixivpy-token/commit/bccbc4f7262855f034b881dc303b0e7488d813fb)
+  - fix: readme
+- [c7e4dac0a536294adf0a6f494cbd45ae9246d827](https://github.com/eggplants/get-pixivpy-token/commit/c7e4dac0a536294adf0a6f494cbd45ae9246d827)
+  - add: badge
+- [40eb89ca2d80001f14e3a73b868793c58312e765](https://github.com/eggplants/get-pixivpy-token/commit/40eb89ca2d80001f14e3a73b868793c58312e765)
+  - add: dockerfile
+- [97bab8bf980ef1c152e7384f224e56ade4272187](https://github.com/eggplants/get-pixivpy-token/commit/97bab8bf980ef1c152e7384f224e56ade4272187)
+  - add: dockerfile
+
+## [v1.0](https://github.com/eggplants/get-pixivpy-token/releases/tag/v1.0) - 2021-09-05 03:49:42 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v0.5...v1.0))
+
+- [8ac6b0ddadec3f62d47745e8e9fdbae0e665dff9](https://github.com/eggplants/get-pixivpy-token/commit/8ac6b0ddadec3f62d47745e8e9fdbae0e665dff9)
+  - v1.0
+- [706fe9753924edc512b92f81d567a9e3c335fecd](https://github.com/eggplants/get-pixivpy-token/commit/706fe9753924edc512b92f81d567a9e3c335fecd)
+  - fix
+- [5be5f4dce8900dc74d959769147c9b7d74d3803e](https://github.com/eggplants/get-pixivpy-token/commit/5be5f4dce8900dc74d959769147c9b7d74d3803e)
+  - fix: readme
+
 ## [v0.5](https://github.com/eggplants/get-pixivpy-token/releases/tag/v0.5) - 2021-08-29 14:28:12 ([compare](https://github.com/eggplants/get-pixivpy-token/compare/v0.4...v0.5))
 
 - [b4fa365dd44d9a8288a61531cdc3c28a4e276881](https://github.com/eggplants/get-pixivpy-token/commit/b4fa365dd44d9a8288a61531cdc3c28a4e276881)